requirements.txt
matplotlib==2.2.3
numpy==1.15.1
tensorflow==1.12.0
pillow
tqdm
scipy